+++ CHANGES.rst
@@ -9,6 +9,10 @@
 
 - Add support for Python 3.7, 3.8 and 3.9.
 
+- Prevent a ``DeprecationWarning`` in Python 3.8+ when using a parameter for
+  ``iso8601_date``, ``rfc850_date``, or ``rfc1123_date`` which has to be
+  converted via its ``__int__`` method.
+
 
 4.2.0 (2017-08-14)
 ==================